+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.12.0] - 2021-03-13
+
+### Added
+
+-   Moving average and adjust logic to utilize it
+
+### Updated
+
+-   buy & sell to utilize an owned boolean
+-   live poc to no longer log theory data
+
 ## [0.11.2] - 2021-03-08
 
 ### Updated
@@ -224,6 +235,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Working demo and API connections
 
+[0.12.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.11.2...v0.12.0
 [0.11.2]: https://github.com/JesseChamberlain/binance-trader/compare/v0.11.1...v0.11.2
 [0.11.1]: https://github.com/JesseChamberlain/binance-trader/compare/v0.11.0...v0.11.1
 [0.11.0]: https://github.com/JesseChamberlain/binance-trader/compare/v0.10.0...v0.11.0